Simple refactor setup network storage from setup script to separate module
