+++ ansible/roles/slurm/tasks/main.yml
@@ -1,5 +1,6 @@
 ---
 # Copyright (C) SchedMD LLC.
+# Copyright 2024 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -67,6 +68,7 @@
   - conf.py
   - resume.py
   - setup.py
+  - setup_network_storage.py
   - startup.sh
   - slurmsync.py
   - suspend.py
